backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
bcrypt==4.2.0
PyJWT==2.9.0
